scripts/compress-images/requirements.txt
black==23.3.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
Pygments==2.14.0
rich==12.6.0
shellingham==1.5.0.post1
typer==0.7.0
